Add files via upload
Currently the pygame tutorial opens a window in pygame and has a function for what occurs each frame and a function to render what has changed. The SWParticles is for social web particles and contains a particle class that currently is just a ton of balls that collide and bounce off of the walls and are effected by gravity. 

To simulate motion what is being done is the Y axis is positive in the negative direction, x is positive to the right, and theta is running clockwise (I am pretty sure that is the case with theta. I think this makes cross products positive if they go clockwise?